all done, constraint for # of clues seems to be broken though
