+++ GrailDiary/GrailDiary/POI.swift
@@ -1,0 +1,9 @@
+//
+//  POI.swift
+//  GrailDiary
+//
+//  Created by John Thomas on 1/23/20.
+//  Copyright © 2020 John Thomas. All rights reserved.
+//
+
+import Foundation
