+++ GrailDiary/GrailDiary/POIDetailViewController.swift
@@ -1,0 +1,30 @@
+//
+//  POIDetailViewController.swift
+//  GrailDiary
+//
+//  Created by John Thomas on 1/23/20.
+//  Copyright © 2020 John Thomas. All rights reserved.
+//
+
+import UIKit
+
+class POIDetailViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
