+++ GrailDiary/GrailDiary/POITableViewCell.swift
@@ -1,0 +1,24 @@
+//
+//  POITableTableViewCell.swift
+//  GrailDiary
+//
+//  Created by John Thomas on 1/23/20.
+//  Copyright © 2020 John Thomas. All rights reserved.
+//
+
+import UIKit
+
+class POITableViewCell: UITableViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
